requirements.txt
yfinance==0.2.18
python-telegram-bot==20.3
PyYAML==6.0